requirements.txt
behave
eventsourcing
flask
flask-restful
pyhamcrest
pytest-bdd
